Move common functions and data types to internal module.
