+++ hgeometry-combinatorial/hgeometry-combinatorial.cabal
@@ -116,7 +116,7 @@ library
 
                     Data.Yaml.Util
 
-  other-modules:
+  other-modules:    Data.PlanarGraph.Internal
 
   -- other-extensions:
   build-depends:
