+++ hgeometry-combinatorial/src/Data/PlanarGraph/Mutable.hs
@@ -1,6 +1,6 @@
 module Data.PlanarGraph.Mutable
   ( -- * Planar graphs
-    PlanarGraph(..)
+    PlanarGraph
   , pgFromFaces   -- :: [[VertexId]] -> ST s (PlanarGraph s)
   , pgFromFacesCV -- :: [CircularVector VertexId] -> ST s (PlanarGraph s)
   , pgClone       -- :: PlanarGraph s -> ST s (PlanarGraph s)
@@ -72,73 +72,27 @@ module Data.PlanarGraph.Mutable
   )
   where
 
-import           Control.Monad
-import           Control.Monad.ST
-import           Data.Bits
-import qualified Data.HashMap.Strict  as HM
-import           Data.STRef
-import qualified Data.Vector          as V (unsafeFreeze)
-import           Data.Vector.Circular (CircularVector)
-import qualified Data.Vector.Circular as CV
-import           Data.Vector.Mutable  (STVector)
-import qualified Data.Vector.Mutable  as V
-import Data.Hashable
-
-import           Data.Coerce
-import           Data.Proxy
-import qualified Data.Vector   as Vector
-import           Linear.Matrix (luSolve)
-import           Linear.V
+import           Control.Monad             (forM_, unless, when)
+import           Control.Monad.ST          (ST)
+import           Data.Bits                 (Bits (xor))
+import           Data.Coerce               (coerce)
+import qualified Data.HashMap.Strict       as HM
+import           Data.Hashable             (Hashable (hashWithSalt))
+import           Data.PlanarGraph.Internal
+import           Data.Proxy                (Proxy (..))
+import           Data.STRef                (modifySTRef', newSTRef, readSTRef, writeSTRef)
+import qualified Data.Vector               as Vector
+import           Data.Vector.Circular      (CircularVector)
+import qualified Data.Vector.Circular      as CV
+import qualified Data.Vector.Mutable       as V
+import           Linear.Matrix             (luSolve)
+import           Linear.V                  (Dim, V (..), reifyDim)
 
 import Debug.Trace
 
 -------------------------------------------------------------------------------
--- Resizeable vector
-
-type GrowVector s v = STRef s (STVector s v)
-
-newVector :: Int -> ST s (GrowVector s v)
-newVector n = newSTRef =<< V.new n
-
-setVector :: GrowVector s v -> v -> ST s ()
-setVector ref val = do
-  vec <- readSTRef ref
-  V.set vec val
-
-readVector :: GrowVector s v -> Int -> ST s v
-readVector ref idx = do
-  v <- readSTRef ref
-  V.read v idx
-
-writeVector :: GrowVector s v -> Int -> v -> ST s ()
-writeVector ref idx val = do
-  v <- readSTRef ref
-  let l = V.length v
-  if idx >= l
-    then {-trace ("Growing: " ++ show (idx, l)) $ -} do
-      v' <- V.grow v ((idx+1)*2)
-      V.write v' idx val
-      writeSTRef ref v'
-    else -- trace ("Writing: " ++ show (idx, l)) $
-      V.write v idx val
-
-freezeVector :: GrowVector s v -> ST s (Vector.Vector v)
-freezeVector ref = Vector.freeze =<< readSTRef ref
-
-thawVector :: Vector.Vector v -> ST s (GrowVector s v)
-thawVector v = newSTRef =<< Vector.thaw v
-
-freezeCircularVector :: Int -> GrowVector s v -> ST s (CircularVector v)
-freezeCircularVector n ref =
-  (CV.unsafeFromVector . Vector.take n) <$> (V.unsafeFreeze =<< readSTRef ref)
-
-
--------------------------------------------------------------------------------
 -- Elements: Half-edges, vertices, faces.
 
-
-
-type HalfEdgeId = Int
 data HalfEdge s = HalfEdge HalfEdgeId (PlanarGraph s)
   deriving Eq
 instance Show (HalfEdge s) where
@@ -150,7 +104,6 @@ instance Hashable (HalfEdge s) where
 data Edge s = Edge Int (PlanarGraph s)
   deriving Eq
 
-type VertexId = Int
 data Vertex s = Vertex VertexId (PlanarGraph s)
   deriving Eq
 instance Show (Vertex s) where
@@ -158,37 +111,14 @@ instance Show (Vertex s) where
 instance Hashable (Vertex s) where
   hashWithSalt salt (Vertex vId _) = hashWithSalt salt vId
 
-type FaceId = Int
 data Face s = Face FaceId (PlanarGraph s) | Boundary FaceId (PlanarGraph s)
   deriving Eq
 instance Show (Face s) where
-  showsPrec d (Face fId _) = showString "Face " . shows fId
+  showsPrec d (Face fId _)     = showString "Face " . shows fId
   showsPrec d (Boundary fId _) = showString "Boundary " . shows fId
 
 -------------------------------------------------------------------------------
 -- Planar graph
-
--- PlanarGraphs have vertices, edges, and faces.
--- Invariant: The half-edge of a boundary vertex is interior, twin is exterior.
-
--- FIXME: Use STRefU ?
--- PlanarGraph with 0 vertices: No edges, no vertices, no faces.
--- PlanarGraph with 1 vertex: No edges, no interior faces.
--- PlanarGraph with 2 vertices: One edge, no interior faces.
--- PlanarGraph with 3+ vertices: Usual properties hold.
-data PlanarGraph s = PlanarGraph
-  { pgNextHalfEdgeId :: !(STRef s HalfEdgeId)
-  , pgNextVertexId   :: !(STRef s VertexId)
-  , pgNextFaceId     :: !(STRef s FaceId)
-  , pgNextBoundaryId :: !(STRef s FaceId)
-  , pgHalfEdgeNext   :: !(GrowVector s HalfEdgeId) -- HalfEdge indexed
-  , pgHalfEdgePrev   :: !(GrowVector s HalfEdgeId) -- HalfEdge indexed
-  , pgHalfEdgeVertex :: !(GrowVector s VertexId)   -- HalfEdge indexed
-  , pgHalfEdgeFace   :: !(GrowVector s FaceId)     -- HalfEdge indexed
-  , pgVertices       :: !(GrowVector s HalfEdgeId) -- Vertex indexed
-  , pgFaces          :: !(GrowVector s HalfEdgeId) -- Face indexed
-  , pgBoundaries     :: !(GrowVector s HalfEdgeId) -- Boundary faces
-  } deriving Eq
 
 panic :: String -> String -> a
 panic tag msg = error $ "Data.PlanarGraph.Mutable." ++ tag ++ ": " ++ msg
@@ -292,7 +222,7 @@ pgFromFacesCV faces = do
   --   If face is invalid:
   --     Find loop and add it as a boundary.
   forM_ (map (`halfEdgeFromId` pg) [0..maxHalfEdgeId-1]) $ \he -> {- trace ("Scan halfedge: " ++ show he) $ -} do
-    f <- halfEdgeFace he
+    -- f <- halfEdgeFace he
     validFace <- faceIsValid <$> halfEdgeFace he
     unless validFace $ {- trace ("Found invalid face: " ++ show f) $ -} do
       face <- faceNewBoundary pg
@@ -565,7 +495,7 @@ faceIsValid = not . faceIsInvalid
 
 -- | O(1)
 faceIsInvalid :: Face s -> Bool
-faceIsInvalid (Face fId _) = fId == maxBound
+faceIsInvalid (Face fId _)     = fId == maxBound
 faceIsInvalid (Boundary fId _) = fId == maxBound
 
 -- | O(1)
@@ -575,7 +505,7 @@ faceFromId fId pg = Face fId pg
 
 -- | O(1)
 faceToId :: Face s -> FaceId
-faceToId (Face fId _) = fId
+faceToId (Face fId _)     = fId
 faceToId (Boundary fId _) = negate fId - 1
 
 -- | O(1)
